完成了公告消息的群体选择、发布、删除、修改

+++ src/main/java/com/scnujxjy/backendpoint/util/excelListener/TeacherInformationListener.java
@@ -157,7 +157,7 @@ public class TeacherInformationListener extends AnalysisEventListener<TeacherInf
 
 
         return TeacherInformationPO.builder()
-                .userId(vo.getUserId())
+                .userId((long) vo.getUserId())
                 .name(vo.getName())
                 .gender(vo.getGender())
                 .birthDate(birthDate)
