+++ src/test/java/com/scnujxjy/backendpoint/TeacherInformationTest/TeacherInformationListener.java
@@ -128,7 +128,7 @@ public class TeacherInformationListener extends AnalysisEventListener<TeacherInf
 
 
         return TeacherInformationPO.builder()
-                .userId(vo.getUserId())
+                .userId((long) vo.getUserId())
                 .name(vo.getName())
                 .gender(vo.getGender())
                 .birthDate(birthDate)
